% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chinook}
\alias{chinook}
\title{Data from Satterthwaite, W.H., Ciancio, J., Crandall, E., Palmer-Zwahlen,
M.L., Grover, A.M., O’Farrell, M.R., Anson, E.C., Mohr, M.S. & Garza,
J.C. (2015). Stock composition and ocean spatial distribution from
California recreational chinook salmon fisheries using genetic stock
identification. Fisheries Research, 170, 166–178. The data
genetic data collected from port-based sampling of recreationally-landed
Chinook salmon in California from 1998-2002.}
\format{
A data frame.
}
\usage{
chinook
}
\description{
Data from Satterthwaite, W.H., Ciancio, J., Crandall, E., Palmer-Zwahlen,
M.L., Grover, A.M., O’Farrell, M.R., Anson, E.C., Mohr, M.S. & Garza,
J.C. (2015). Stock composition and ocean spatial distribution from
California recreational chinook salmon fisheries using genetic stock
identification. Fisheries Research, 170, 166–178. The data
genetic data collected from port-based sampling of recreationally-landed
Chinook salmon in California from 1998-2002.
}
\keyword{datasets}
