% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coddiet}
\alias{coddiet}
\title{Data from Magnussen, E. 2011. Food and feeding habits of cod (Gadus morhua)
on the Faroe Bank. – ICES Journal of Marine Science, 68: 1909–1917. The data
here are Table 3 from the paper, with sample proportions (columns w) multiplied
by total weight to yield total grams (g) for each sample-diet item combination. Dashes
have been replaced with 0s.}
\format{
A data frame.
}
\usage{
coddiet
}
\description{
Data from Magnussen, E. 2011. Food and feeding habits of cod (Gadus morhua)
on the Faroe Bank. – ICES Journal of Marine Science, 68: 1909–1917. The data
here are Table 3 from the paper, with sample proportions (columns w) multiplied
by total weight to yield total grams (g) for each sample-diet item combination. Dashes
have been replaced with 0s.
}
\keyword{datasets}
