% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_prior.R
\name{fit_dirichlet}
\alias{fit_dirichlet}
\title{Extract point estimates of compositions from fitted model.}
\usage{
fit_dirichlet(data)
}
\arguments{
\item{data}{The data to fit the dirichlet distribution to}
}
\description{
Extract point estimates of compositions from fitted model.
}
