% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{parse_re_formula}
\alias{parse_re_formula}
\title{Fit a trinomial mixture model that optionally includes covariates to estimate
effects of factor or continuous variables on proportions.}
\usage{
parse_re_formula(formula, data)
}
\arguments{
\item{formula}{The model formula for the design matrix.}

\item{data}{The data matrix used to construct RE design matrix}
}
\description{
Fit a trinomial mixture model that optionally includes covariates to estimate
effects of factor or continuous variables on proportions.
}
