% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_prior.R
\name{rmspe_calc}
\alias{rmspe_calc}
\title{Find appropriate prior for a given target distribution.}
\usage{
rmspe_calc(par, n_bins, n_draws, target)
}
\arguments{
\item{par}{The parameter (standard deviation) to be searched over to find a Dirichlet equivalent}

\item{n_bins}{Bins for the Dirichlet distribution}

\item{n_draws}{Numbers of samples to use for doing calculation}

\item{target}{The goal of the specified prior, e.g. 1 or 1/n_bins}
}
\description{
Extract point estimates of compositions from fitted model.
}
